// web/src/api/generated/model/createLibraryInputBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { CreateLibraryInputBodyEncryptionMode } from "./createLibraryInputBodyEncryptionMode.ts";

export interface CreateLibraryInputBody {
  /**
   * @minLength 1
   * @maxLength 64
   */
  backend_id: string;
  encryption_mode: CreateLibraryInputBodyEncryptionMode;
  key_envelope?: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
}